requirements.txt
pandas==1.5.1
MLB-StatsAPI
numpy
crewai_tools==0.4.8
crewai==0.41.1
langchain_groq==0.1.6
fpdf
streamlit